services/games/app/requirements.txt
flask==2.3.2
pyjwt==1.5.0
cryptography==36.0.2